+++ module-apps/application-messages/models/ThreadsSearchResultsModel.hpp
@@ -13,14 +13,13 @@ namespace gui::model
       public:
         ThreadsSearchResultsModel(app::Application *app);
 
-        unsigned int getMinimalItemHeight() const override;
-        ListItem *getItem(Order order) override;
+        auto getMinimalItemHeight() const -> unsigned int override;
+        auto getItem(Order order) -> ListItem * override;
         /// empty, size get in requestRecords
         void requestRecords(uint32_t offset, uint32_t limit) override;
         /// set what we need to search
         void setSearchValue(const UTF8 &search_value);
 
-        // virtual method for db::QueryListener
-        auto handleQueryResponse(db::QueryResult *) -> bool override;
+        auto handleQueryResponse(db::QueryResult *) -> bool;
     };
 }; // namespace gui::model